+++ GNAR/GameManager.m
@@ -35,7 +35,7 @@ static GameManager *gameManager = nil;
 - (void)printCurrentGame
 {
     NSLog(@"Saved singleton game: %@", self.currentGame.name);
-    NSLog(@"Saved singleton game player count: %lu", (unsigned long)self.currentGame.playersArray.count);
+    NSLog(@"Saved singleton game player count: %lu", self.currentGame.playersArray.count);
 }
 
 //- (GameManager *)init
